Update accoring to template is done.
